edit useMemo and delete file unused
