--- src/App.js
@@ -1,5 +1,0 @@
-import "./App.css";
-
-function App() {}
-
-export default App;
