+++ src/Square.js
@@ -1,9 +1,9 @@
 function Square(props) {
-    return (
-        <button className="square" onClick={props.onClick}>
-          {props.value}
-        </button>
-      );
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
 }
 
 export default Square;
