+++ src/components/Board.js
@@ -5,7 +5,7 @@ function Board(props) {
   return (
     <div>
       <h1>Tic Tac Toe</h1>
-      <div className="status">{props.turn}</div>
+      <div className="status">{props.winner}</div>
       <div className="board">
         {props.squares.map((item, index) => {
           return (
