+++ src/components/Game.js
@@ -1,12 +1,12 @@
 import { useMemo, useState } from "react";
 import Board from "./Board";
-import calculateWinner from "./Winner";
 
 function Game(props) {
+  const resetbox = () => {
+    return Array(9).fill("");
+  };
   const [xIsNext, setxIsNext] = useState(true);
-  const [box, setBox] = useState(Array(9).fill(""));
-  const winner = useMemo(() => calculateWinner(box), [xIsNext]);
-  let turn;
+  const [box, setBox] = useState(resetbox);
 
   const handleClick = (i) => {
     if (calculateWinner(box) || box[i]) {
@@ -17,27 +17,50 @@ function Game(props) {
     setxIsNext(!xIsNext);
   };
 
-  if (winner) {
-    turn = winner + " is Winner!";
-  } else if (!winner && box.includes("") === false) {
-    turn = "Drawn!!!";
-  } else {
-    turn = (xIsNext ? "X" : "O") + " Play";
-  }
+  const winner = useMemo(() => {
+    let turn;
+    if (calculateWinner(box)) {
+      return (turn = calculateWinner(box) + " is Winner!");
+    } else if (!calculateWinner(box) && box.includes("") === false) {
+      return (turn = "Drawn!!!");
+    } else {
+      return (turn = (xIsNext ? "X" : "O") + " Play");
+    }
+  }, [xIsNext]);
 
   const handleRestart = () => {
     setxIsNext(true);
-    setBox(Array(9).fill(""));
+    setBox(resetbox);
   };
 
   return (
     <Board
       squares={box}
-      turn={turn}
+      winner={winner}
       onClick={handleClick}
       onRestart={handleRestart}
     />
   );
+}
+
+function calculateWinner(squares) {
+  const lines = [
+    [0, 1, 2],
+    [3, 4, 5],
+    [6, 7, 8],
+    [0, 3, 6],
+    [1, 4, 7],
+    [2, 5, 8],
+    [0, 4, 8],
+    [2, 4, 6],
+  ];
+  for (let i in lines) {
+    const [a, b, c] = lines[i];
+    if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c]) {
+      return squares[a];
+    }
+  }
+  return null;
 }
 
 export default Game;
